Made heap comparator required, fixed type dilemma
There was a potential bug if you created a Heap of type T (where T is not number) and don't pass a comparator. The comparator would default to the (number-based) minHeapComparator, but it could be incompatible with type T. Now the comparator must be passed and its parameter types are enforced by TypeScript. This also means custom comparators won't need to be cast as `any` to be passed to the Heap.